pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vnstock-research"
version = "0.1.0"
description = "Research-friendly Vietnamese stock forecasting repository with shared data contracts and comparable model outputs."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "PyYAML>=6.0",
  "scipy>=1.12",
  "seaborn>=0.13",
  "vnstock>=3.5.1",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.29",
  "nbformat>=5.10",
  "pytest>=8.1",
  "ruff>=0.3",
]
dl = [
  "torch>=2.2",
]
kronos = [
  "einops>=0.8.1",
  "safetensors>=0.6.2",
  "huggingface_hub>=0.33.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
